requirements.txt
paho-mqtt
schedule
requests
python-dotenv
sqlalchemy
mysql-connector-python
apscheduler